+++ packages/design-system/src/views/HomeApp.vue
@@ -72,8 +72,8 @@
     position: relative;
     padding-top: 34%;
     overflow: hidden;
-    background: $p-color-neutral-grey-2;
-    border-bottom: 1px solid $p-color-neutral-grey-2;
+    background: $p-color-theme-dark-neutral-2;
+    border-bottom: 1px solid $p-color-theme-dark-neutral-2;
   }
 
   .video {
@@ -102,16 +102,16 @@
 
     p a {
       text-decoration: none;
-      color: $p-color-porsche-red;
+      color: $p-color-theme-light-brand;
       transition: color $p-animation-hover-duration $p-animation-hover-bezier;
 
       &:hover {
-        color: $p-color-porsche-dark-red;
+        color: darken($p-color-theme-light-brand, 10%);
       }
     }
 
     h2 + p a {
-      border-bottom: 1px solid $p-color-porsche-red;
+      border-bottom: 1px solid $p-color-theme-light-brand;
     }
   }
 </style>